# packages/api/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "SelfFeed API",
		"version": "1.0.0",
		"description": "Stable HTTP API for the SelfFeed web client and future mobile clients."
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		},
		"schemas": {
			"ApiError": {
				"type": "object",
				"required": [
					"error"
				],
				"properties": {
					"error": {
						"type": "object",
						"required": [
							"code",
							"message"
						],
						"properties": {
							"code": {
								"type": "string"
							},
							"message": {
								"type": "string"
							},
							"details": {}
						}
					}
				}
			},
...
			"FeedDiscoveryPresentation": {
				"type": "object",
				"required": [
					"required",
					"candidates"
				],
				"properties": {
					"required": {
						"type": "boolean"
					},
					"candidates": {
						"type": "array",
						"items": {
							"type": "object",
							"required": [
								"id",
								"requestId",
								"url",
								"title",
								"type",
								"expiresAt"
							],
							"properties": {
								"id": {
									"type": "string",
									"format": "uuid"
								},
								"requestId": {
									"type": "string",
									"format": "uuid"
								},
								"url": {
									"type": "string",
									"format": "uri"
								},
								"title": {
									"type": [
										"string",
										"null"
									]
								},
								"type": {
									"type": "string"
								},
								"expiresAt": {
									"type": "string",
									"format": "date-time"
								}
							}
						}
					}
				}
			},
			"FeedRefreshItemStatus": {
				"type": "object",
				"required": [
					"feedId",
					"sourceId",
					"jobId",
					"status",
					"feedTitle",
					"errorCode",
					"errorDetails",
					"nextEligibleAt",
					"publisherRequestStarted",
					"lastFetchAt"
				],
				"properties": {
					"feedId": {
						"type": [
							"string",
							"null"
						],
						"format": "uuid"
					},
					"sourceId": {
						"type": [
							"string",
							"null"
						],
						"format": "uuid"
					},
					"jobId": {
						"type": [
							"string",
							"null"
						],
						"format": "uuid"
					},
					"status": {
						"type": "string"
					},
					"feedTitle": {
						"type": [
							"string",
							"null"
						]
					},
					"errorCode": {
						"type": [
							"string",
							"null"
						]
					},
					"errorDetails": {
						"type": [
							"string",
							"null"
						]
					},
					"nextEligibleAt": {
						"type": [
							"string",
							"null"
						],
						"format": "date-time"
					},
					"publisherRequestStarted": {
						"type": "boolean"
					},
					"lastFetchAt": {
						"type": [
							"string",
							"null"
						],
						"format": "date-time"
					}
				}
			},
			"FeedSyncAllStatus": {
				"type": "object",
				"required": [
					"queued",
					"running",
					"active",
					"stale",
					"queuedAt",
					"startedAt",
					"heartbeatAt",
					"totalFeeds",
					"completedFeeds",
					"syncedFeeds",
					"failedFeeds",
					"skippedFeeds",
					"newArticles",
					"articleRevision",
					"jobId",
					"scope",
					"items"
				],
				"properties": {
					"requestId": {
						"type": [
							"string",
							"null"
						],
						"format": "uuid"
					},
					"status": {
						"type": "string",
						"enum": [
							"pending",
							"running",
							"completed",
							"completed_with_errors"
						]
					},
					"queued": {
						"type": "boolean"
					},
					"running": {
						"type": "boolean"
					},
					"active": {
						"type": "boolean"
					},
					"stale": {
						"type": "boolean"
					},
					"queuedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"startedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"heartbeatAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"totalFeeds": {
						"type": "integer",
						"minimum": 0
					},
					"completedFeeds": {
						"type": "integer",
						"minimum": 0
					},
					"syncedFeeds": {
						"type": "integer",
						"minimum": 0
					},
					"failedFeeds": {
						"type": "integer",
						"minimum": 0
					},
					"skippedFeeds": {
						"type": "integer",
						"minimum": 0
					},
					"pendingFeeds": {
						"type": "integer",
						"minimum": 0
					},
					"runningFeeds": {
						"type": "integer",
						"minimum": 0
					},
					"deadFeeds": {
						"type": "integer",
						"minimum": 0
					},
					"newArticles": {
						"type": "integer",
						"minimum": 0
					},
					"articleRevision": {
						"type": "integer",
						"minimum": 0
					},
					"jobId": {
						"type": [
							"string",
							"null"
						]
					},
					"scope": {
						"type": "object",
						"properties": {
							"feedId": {
								"type": "string",
								"format": "uuid"
							},
							"categoryId": {
								"type": "string",
								"format": "uuid"
							}
						}
					},
					"items": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/FeedRefreshItemStatus"
						}
					}
				}
			},
			"FeedDiscoveryCandidate": {
				"type": "object",
				"required": [
					"id",
					"requestId",
					"candidateUrl",
					"normalizedCandidateUrl",
					"status",
					"expiresAt"
				],
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"requestId": {
						"type": "string",
						"format": "uuid"
					},
					"candidateUrl": {
						"type": "string",
						"format": "uri"
					},
					"normalizedCandidateUrl": {
						"type": "string",
						"format": "uri"
					},
					"title": {
						"type": [
							"string",
							"null"
						]
					},
					"type": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"expiresAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
...
			"Feed": {
				"type": "object",
				"required": [
					"id",
					"userId",
					"categoryId",
					"title",
					"siteUrl",
					"feedUrl",
					"faviconUrl",
					"description",
					"pollingIntervalMinutes",
					"lastSyncedAt",
					"lastSyncError",
					"lastSyncErrorAt",
					"syncStatus",
					"createdAt",
					"updatedAt"
				],
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"userId": {
						"type": "string",
						"format": "uuid"
					},
					"categoryId": {
						"type": "string",
						"format": "uuid"
					},
					"title": {
						"type": "string"
					},
					"siteUrl": {
						"type": [
							"string",
							"null"
						]
					},
					"feedUrl": {
						"type": "string",
						"format": "uri"
					},
					"faviconUrl": {
						"type": [
							"string",
							"null"
						]
					},
					"description": {
						"type": [
							"string",
							"null"
						]
					},
					"pollingIntervalMinutes": {
						"type": "integer"
					},
					"lastSyncedAt": {
						"type": [
							"string",
							"null"
						],
						"format": "date-time"
					},
					"lastSyncError": {
						"type": [
							"string",
							"null"
						]
					},
					"lastSyncErrorAt": {
						"type": [
							"string",
							"null"
						],
						"format": "date-time"
					},
					"nextSyncAt": {
						"type": "string",
						"format": "date-time"
					},
					"syncStatus": {
						"type": "string",
						"enum": [
							"idle",
							"syncing",
							"error",
							"pending",
							"backoff",
							"paused",
							"discovery_required",
							"replacement_pending"
						]
					},
					"lifecycleStatus": {
						"type": "string",
						"enum": [
							"pending",
							"active",
							"backoff",
							"paused",
							"error",
							"discovery_required",
							"replacement_pending"
						]
					},
					"sourceId": {
						"type": [
							"string",
							"null"
						],
						"format": "uuid"
					},
					"pendingSourceId": {
						"type": [
							"string",
							"null"
						],
						"format": "uuid"
					},
					"pendingFeedUrl": {
						"type": [
							"string",
							"null"
						],
						"format": "uri"
					},
					"sourceState": {
						"type": [
							"string",
							"null"
						]
					},
					"sourceErrorCode": {
						"type": [
							"string",
							"null"
						]
					},
					"sourceErrorDetails": {
						"type": [
							"string",
							"null"
						]
					},
					"lastFetchAt": {
						"type": [
							"string",
							"null"
						],
						"format": "date-time"
					},
					"lastSuccessAt": {
						"type": [
							"string",
							"null"
						],
						"format": "date-time"
					},
					"nextEligibleFetchAt": {
						"type": [
							"string",
							"null"
						],
						"format": "date-time"
					},
					"replacementRequestedAt": {
						"type": [
							"string",
							"null"
						],
						"format": "date-time"
					},
					"discovery": {
						"$ref": "#/components/schemas/FeedDiscoveryPresentation"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"FeedWithCounts": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Feed"
					},
					{
						"type": "object",
						"required": [
							"unreadCount"
						],
						"properties": {
							"unreadCount": {
								"type": "integer"
							}
						}
					}
				]
			},
...
			"SyncRun": {
				"type": "object",
				"required": [
					"id",
					"feedId",
					"feedTitle",
					"startedAt",
					"finishedAt",
					"status",
					"httpStatus",
					"itemCount",
					"errorMessage"
				],
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"feedId": {
						"type": "string",
						"format": "uuid"
					},
					"feedTitle": {
						"type": [
							"string",
							"null"
						]
					},
					"startedAt": {
						"type": "string",
						"format": "date-time"
					},
					"finishedAt": {
						"type": [
							"string",
							"null"
						],
						"format": "date-time"
					},
					"status": {
						"type": "string",
						"enum": [
							"running",
							"success",
							"failed"
						]
					},
					"httpStatus": {
						"type": [
							"integer",
							"null"
						]
					},
					"itemCount": {
						"type": "integer"
					},
					"errorMessage": {
						"type": [
							"string",
							"null"
						],
						"maxLength": 500
					}
				}
			},
			"OpmlImportWarning": {
				"type": "object",
				"required": [
					"code",
					"message"
				],
				"properties": {
					"code": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"feedUrl": {
						"type": "string"
					},
					"categoryPath": {
						"type": "array",
						"items": {
							"type": "string"
						}
					}
				}
			},
			"OpmlImportSummary": {
				"type": "object",
				"required": [
					"createdCategories",
					"createdFeeds",
					"skippedDuplicates",
					"invalidEntries",
					"warnings"
				],
				"properties": {
					"createdCategories": {
						"type": "integer"
					},
					"createdFeeds": {
						"type": "integer"
					},
					"skippedDuplicates": {
						"type": "integer"
					},
					"invalidEntries": {
						"type": "integer"
					},
					"warnings": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/OpmlImportWarning"
						}
					}
				}
			}
		}
	},
	"paths": {
...
		"/feeds/{feedId}/replacement/cancel": {
			"post": {
				"tags": [
					"Feeds"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "feedId",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": [
										"data"
									],
									"properties": {
										"data": {
											"$ref": "#/components/schemas/FeedWithCounts"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/feeds/discovery/{requestId}": {
			"get": {
				"tags": [
					"Feeds"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "requestId",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/FeedDiscoveryCandidate"
									}
								}
							}
						}
					}
				}
			}
		},
		"/feeds/discovery/candidates/{candidateId}/select": {
			"post": {
				"tags": [
					"Feeds"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "candidateId",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"202": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								}
							}
						}
					}
				}
			}
		},
		"/feeds/{feedId}/sync-runs": {
			"get": {
				"tags": [
					"Feeds"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "feedId",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"schema": {
							"type": "integer",
							"minimum": 1,
							"maximum": 100,
							"default": 25
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"schema": {
							"type": "string",
							"pattern": "^\\d+$"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": [
										"data"
									],
									"properties": {
										"data": {
											"type": "object",
											"required": [
												"runs",
												"cursor",
												"hasMore"
											],
											"properties": {
												"runs": {
													"type": "array",
													"items": {
														"$ref": "#/components/schemas/SyncRun"
													}
												},
												"cursor": {
													"type": [
														"string",
														"null"
													]
												},
												"hasMore": {
													"type": "boolean"
												}
											}
										}
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/feeds": {
			"get": {
				"tags": [
					"Feeds"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": [
										"data"
									],
									"properties": {
										"data": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/FeedWithCounts"
											}
										}
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": [
					"Feeds"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": [
									"categoryId",
									"feedUrl"
								],
								"properties": {
									"categoryId": {
										"type": "string",
										"format": "uuid"
									},
									"feedUrl": {
										"type": "string",
										"format": "uri",
										"maxLength": 2048
									},
									"title": {
										"type": "string",
										"minLength": 1,
										"maxLength": 255
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": [
										"data"
									],
									"properties": {
										"data": {
											"$ref": "#/components/schemas/FeedWithCounts"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/feeds/import/opml": {
			"post": {
				"tags": [
					"Feeds"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"required": [
									"file"
								],
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": [
										"data"
									],
									"properties": {
										"data": {
											"$ref": "#/components/schemas/OpmlImportSummary"
										}
									}
								}
							}
						}
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
		"/feeds/export/opml": {
			"get": {
				"tags": [
					"Feeds"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "OPML subscription export",
						"content": {
							"application/xml": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				}
			}
		},
		"/feeds/{feedId}": {
			"patch": {
				"tags": [
					"Feeds"
				],
				"description": "Non-URL edits apply immediately. In v2 a URL change validates asynchronously while the old source remains usable; successful validation atomically removes old articles/read rows and activates the replacement.",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "feedId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"categoryId": {
										"type": "string",
										"format": "uuid"
									},
									"feedUrl": {
										"type": "string",
										"format": "uri",
										"maxLength": 2048
									},
									"title": {
										"type": "string",
										"minLength": 1,
										"maxLength": 255
									},
									"pollingIntervalMinutes": {
										"type": "integer",
										"minimum": 5,
										"maximum": 1440
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": [
										"data"
									],
									"properties": {
										"data": {
											"$ref": "#/components/schemas/FeedWithCounts"
										}
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": [
					"Feeds"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "feedId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								}
							}
						}
					}
				}
			}
		},
		"/feeds/sync": {
			"post": {
				"tags": [
					"Feeds"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "Idempotency-Key",
						"schema": {
							"type": "string",
							"maxLength": 255
						}
					},
					{
						"in": "query",
						"name": "feedId",
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"in": "query",
						"name": "categoryId",
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"202": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								}
							}
						}
					}
				}
			}
		},
		"/feeds/sync/status": {
			"get": {
				"tags": [
					"Feeds"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "requestId",
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": [
										"data"
									],
									"properties": {
										"data": {
											"$ref": "#/components/schemas/FeedSyncAllStatus"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/feeds/{feedId}/sync": {
			"post": {
				"tags": [
					"Feeds"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "feedId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								}
							}
						}
					},
					"202": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								}
							}
						}
					}
				}
			}
		},
...
	}
}